refactor(store): ActionType, ActionCreator, Operation in @store/data divided by functions (module7-task1)
