+++ package.json
@@ -52,6 +52,7 @@
   },
   "dependencies": {
     "axios": "0.19.2",
+    "classnames": "2.2.6",
     "nanoid": "3.1.10",
     "prop-types": "15.7.2",
     "react": "16.13.1",
