+++ src/store/data/action-type.js
@@ -1,0 +1,11 @@
+const ActionType = {
+  LOAD_FILMS: `LOAD_FILMS`,
+  CHOOSE_GENRE: `CHOOSE_GENRE`,
+  CHOOSE_FILM_ID: `CHOOSE_FILM_ID`,
+  LOAD_PROMO_FILM: `LOAD_PROMO_FILM`,
+  LOAD_FILM_REVIEWS: `LOAD_FILM_REVIEWS`,
+  ADD_ERROR_NOTIFICATION: `ADD_ERROR_NOTIFICATION`,
+  REMOVE_NOTIFICATION: `REMOVE_NOTIFICATION`,
+};
+
+export default ActionType;
